.devcontainer/requirements.txt
discord.py==1.7.3
RPi.GPIO==0.7.0
